Bandpass Variance for Long-Memory Processes

Computes the band-pass variance for fractional difference (FD) or seasonal persistent (SP) processes using numeric integration of their spectral density function.

bandpass.fdp(a, b, d)

bandpass.spp(a, b, d, fG)

bandpass.spp2(a, b, d1, f1, d2, f2)

bandpass.var.spp(delta, fG, J, Basis, Length)

Arguments

fG, f1, f2

Gegenbauer frequency.

J

Depth of the wavelet transform.

Basis

Logical vector representing the adaptive basis.

Length

Number of elements in Basis.

a

Left-hand boundary for the definite integral.

b

Right-hand boundary for the definite integral.

d, delta, d1, d2

Fractional difference parameter.

Value

Band-pass variance for the FD or SP process between \(a\) and \(b\).

Details

See references.

References

McCoy, E. J., and A. T. Walden (1996) Wavelet analysis and synthesis of stationary long-memory processes, Journal for Computational and Graphical Statistics, 5, No. 1, 26-56.

Whitcher, B. (2001) Simulating Gaussian stationary processes with unbounded spectra, Journal for Computational and Graphical Statistics, 10, No. 1, 112-134.

Author

B. Whitcher