(Inverse) Discrete Wavelet Packet Transforms in Two Dimensions

All possible filtering combinations (low- and high-pass) are performed to decompose a matrix or image. The resulting coefficients are associated with a quad-tree structure corresponding to a partitioning of the two-dimensional frequency plane.

dwpt.2d(x, wf="la8", J=4, boundary="periodic")
idwpt.2d(y, y.basis)
<!-- %modwpt.2d(x, wf="la8", J=4, boundary="periodic") -->

Arguments

x
a matrix or image containing the data be to decomposed. This ojbect must be dyadic (power of 2) in length in each dimension.

wf
Name of the wavelet filter to use in the decomposition. By default this is set to "la8", the Daubechies orthonormal compactly supported wavelet of length \(L=8\) (Daubechies, 1992), least asymmetric family.

J
Specifies the depth of the decomposition. This must be a number less than or equal to \(\log(\mbox{length}(x),2)\).

boundary
Character string specifying the boundary condition. If boundary=="periodic" the default, then the vector you decompose is assumed to be periodic on its defined interval,
if boundary=="reflection", the vector beyond its boundaries is assumed to be a symmetric reflection of itself.

y
dwpt.2d object (list-based structure of matrices)

y.basis
Boolean vector, the same length as \(y\), where TRUE means the basis tensor should be used in the reconstruction.

Value

Basically, a list with the following components

w?.?-w?.?

Wavelet coefficient matrices (images). The first index is associated with the scale of the decomposition while the second is associated with the frequency partition within that level. The left and right strings, separated by the dash `-', correspond to the first \((x)\) and second \((y)\) dimensions.

wavelet

Name of the wavelet filter used.

boundary

How the boundaries were handled.

Details

The code implements the two-dimensional DWPT using the pyramid algorithm of Mallat (1989).

References

Mallat, S. G. (1989) A theory for multiresolution signal decomposition: the wavelet representation, IEEE Transactions on Pattern Analysis and Machine Intelligence, 11, No. 7, 674-693.

Wickerhauser, M. V. (1994) Adapted Wavelet Analysis from Theory to Software, A K Peters.

See also

Author

B. Whitcher