Discrete Hilbert Wavelet Transforms

The discrete Hilbert wavelet transforms (DHWTs) for seasonal and time-varying time series analysis. Transforms include the usual orthogonal (decimated), maximal-overlap (non-decimated) and maximal-overlap packet transforms.

dwt.hilbert(x, wf, n.levels=4, boundary="periodic", ...)
dwt.hilbert.nondyadic(x, ...)
idwt.hilbert(y)
modwt.hilbert(x, wf, n.levels=4, boundary="periodic", ...)
imodwt.hilbert(y)
modwpt.hilbert(x, wf, n.levels=4, boundary="periodic")

Arguments

x
Real-valued time series or vector of observations.

wf
Hilbert wavelet pair

n.levels
Number of levels (depth) of the wavelet transform.

boundary
Boundary treatment, currently only periodic and reflection.

y
Hilbert wavelet transform object (list).

...
Additional parametes to be passed on.

References

Selesnick, I. (200X). IEEE Signal Processing Magazine

Selesnick, I. (200X). IEEE Transactions in Signal Processing

Whither, B. and P.F. Craigmile (2004). Multivariate Spectral Analysis Using Hilbert Wavelet Pairs, International Journal of Wavelets, Multiresolution and Information Processing, to appear.

See also

Author

B. Whitcher