Select a Hilbert Wavelet Pair

Converts name of Hilbert wavelet pair to filter coefficients.

hilbert.filter(name)

Arguments

name
Character string of Hilbert wavelet pair, see acceptable names below (e.g., "k3l3").

Details

Simple switch statement selects the appropriate HWP. There are two parameters that define a Hilbert wavelet pair using the notation of Selesnick (2001,2002), \(K\) and \(L\). Currently, the only implemented combinations \((K,L)\) are (3,3), (3,5), (4,2) and (4,4).

Value

List containing the following items:

L

length of the wavelet filter

h0,g0

low-pass filter coefficients

h1,g1

high-pass filter coefficients

References

Selesnick, I.W. (2001). Hilbert transform pairs of wavelet bases. IEEE Signal Processing Letters\/~8(6), 170--173.

Selesnick, I.W. (2002). The design of approximate Hilbert transform pairs of wavelet bases. IEEE Transactions on Signal Processing\/~50(5), 1144--1152.

See also

Author

B. Whitcher

Examples

hilbert.filter("k3l3")
#> $length
#> [1] 12
#> 
#> $hpf
#> $hpf[[1]]
#>  [1] -0.0022260892  0.0426791770  0.0248291600 -0.4982782400  0.7997265200
#>  [6] -0.2867863600 -0.1564275500  0.0331898960  0.0434276420  0.0022046914
#> [11] -0.0022229002 -0.0001159435
#> 
#> $hpf[[2]]
#>  [1] -1.558262e-02  4.943225e-02  2.167541e-01 -7.458501e-01  6.133371e-01
#>  [6]  1.550640e-02 -1.270504e-01 -3.236969e-02  1.970114e-02  6.190912e-03
#> [11] -5.254341e-05 -1.656336e-05
#> 
#> 
#> $lpf
#> $lpf[[1]]
#>  [1]  0.0001159435 -0.0022229002 -0.0022046914  0.0434276420 -0.0331898960
#>  [6] -0.1564275500  0.2867863600  0.7997265200  0.4982782400  0.0248291600
#> [11] -0.0426791770 -0.0022260892
#> 
#> $lpf[[2]]
#>  [1]  1.656336e-05 -5.254341e-05 -6.190912e-03  1.970114e-02  3.236969e-02
#>  [6] -1.270504e-01 -1.550640e-02  6.133371e-01  7.458501e-01  2.167541e-01
#> [11] -4.943225e-02 -1.558262e-02
#> 
#> 
hilbert.filter("k3l5")
#> $length
#> [1] 12
#> 
#> $hpf
#> $hpf[[1]]
#>  [1] -5.854176e-06  2.299268e-04  2.864101e-03 -1.273398e-02 -5.957379e-02
#>  [6]  1.300891e-01  2.653746e-01 -7.875716e-01  5.340248e-01 -3.981034e-03
#> [11] -4.557677e-02 -3.574788e-02  1.021288e-02  2.614091e-03 -2.131052e-04
#> [16] -5.425879e-06
#> 
#> $hpf[[2]]
#>  [1] -6.439594e-05 -4.664223e-05  9.547891e-03  6.282690e-03 -1.233684e-01
#>  [6]  4.621801e-03  5.832856e-01 -7.650914e-01  2.292948e-01  7.456225e-02
#> [11]  9.250975e-03 -2.860147e-02 -8.400312e-04  1.166470e-03  3.572714e-07
#> [16] -4.932617e-07
#> 
#> 
#> $lpf
#> $lpf[[1]]
#>  [1]  5.425879e-06 -2.131052e-04 -2.614091e-03  1.021288e-02  3.574788e-02
#>  [6] -4.557677e-02  3.981034e-03  5.340248e-01  7.875716e-01  2.653746e-01
#> [11] -1.300891e-01 -5.957379e-02  1.273398e-02  2.864101e-03 -2.299268e-04
#> [16] -5.854176e-06
#> 
#> $lpf[[2]]
#>  [1]  4.932617e-07  3.572714e-07 -1.166470e-03 -8.400312e-04  2.860147e-02
#>  [6]  9.250975e-03 -7.456225e-02  2.292948e-01  7.650914e-01  5.832856e-01
#> [11] -4.621801e-03 -1.233684e-01 -6.282690e-03  9.547891e-03  4.664223e-05
#> [16] -6.439594e-05
#> 
#> 
hilbert.filter("k4l2")
#> $length
#> [1] 12
#> 
#> $hpf
#> $hpf[[1]]
#>  [1]  0.002285229 -0.017099408 -0.061694251  0.160409270  0.227520750
#>  [6] -0.774586170  0.560358370 -0.041525062 -0.034722190 -0.036090743
#> [11]  0.013358873  0.001785330
#> 
#> $hpf[[2]]
#>  [1]  0.0114261460  0.0059121296 -0.1332013800  0.0403150080  0.5409737900
#>  [6] -0.7795662200  0.2746430800  0.0584667250  0.0134499020 -0.0325914860
#> [11] -0.0001847535  0.0003570660
#> 
#> 
#> $lpf
#> $lpf[[1]]
#>  [1] -0.001785330  0.013358873  0.036090743 -0.034722190  0.041525062
#>  [6]  0.560358370  0.774586170  0.227520750 -0.160409270 -0.061694251
#> [11]  0.017099408  0.002285229
#> 
#> $lpf[[2]]
#>  [1] -0.0003570660 -0.0001847535  0.0325914860  0.0134499020 -0.0584667250
#>  [6]  0.2746430800  0.7795662200  0.5409737900 -0.0403150080 -0.1332013800
#> [11] -0.0059121296  0.0114261460
#> 
#> 
hilbert.filter("k4l4")
#> $length
#> [1] 16
#> 
#> $hpf
#> $hpf[[1]]
#>  [1] -2.593319e-05  6.742522e-04  5.732357e-03 -1.697939e-02 -6.975951e-02
#>  [6]  1.337267e-01  2.790955e-01 -7.833091e-01  5.302173e-01 -8.136445e-03
#> [11] -5.068726e-02 -3.860533e-02  1.320347e-02  5.548244e-03 -6.690907e-04
#> [16] -2.573467e-05
#> 
#> $hpf[[2]]
#>  [1] -2.333987e-04 -1.556956e-04  1.548964e-02  9.196125e-03 -1.352010e-01
#>  [6] -5.170879e-03  5.883470e-01 -7.574938e-01  2.329811e-01  7.708457e-02
#> [11]  7.702384e-03 -3.355466e-02 -1.980900e-03  2.990384e-03  1.907454e-06
#> [16] -2.859407e-06
#> 
#> 
#> $lpf
#> $lpf[[1]]
#>  [1]  2.573467e-05 -6.690907e-04 -5.548244e-03  1.320347e-02  3.860533e-02
#>  [6] -5.068726e-02  8.136445e-03  5.302173e-01  7.833091e-01  2.790955e-01
#> [11] -1.337267e-01 -6.975951e-02  1.697939e-02  5.732357e-03 -6.742522e-04
#> [16] -2.593319e-05
#> 
#> $lpf[[2]]
#>  [1]  2.859407e-06  1.907454e-06 -2.990384e-03 -1.980900e-03  3.355466e-02
#>  [6]  7.702384e-03 -7.708457e-02  2.329811e-01  7.574938e-01  5.883470e-01
#> [11]  5.170879e-03 -1.352010e-01 -9.196125e-03  1.548964e-02  1.556956e-04
#> [16] -2.333987e-04
#> 
#>