Wavelet Shrinkage via Thresholding

Perform wavelet shrinkage using data-analytic, hybrid SURE, manual, SURE, or universal thresholding.

da.thresh(wc, alpha = .05, max.level = 4, verbose = FALSE, return.thresh = FALSE)

hybrid.thresh(wc, max.level = 4, verbose = FALSE, seed = 0)

manual.thresh(wc, max.level = 4, value, hard = TRUE)

sure.thresh(wc, max.level = 4, hard = TRUE)

universal.thresh(wc, max.level = 4, hard = TRUE)

universal.thresh.modwt(wc, max.level = 4, hard = TRUE)

Arguments

wc

wavelet coefficients

max.level

maximum level of coefficients to be affected by threshold

value

threshold value (only utilized in manual.thresh)

hard

Boolean value, if hard=F then soft thresholding is used

alpha

level of the hypothesis tests

verbose

if verbose=TRUE then information is printed to the screen

seed

sets random seed (only utilized in hybrid.thresh)

return.thresh

if return.thresh=TRUE then the vector of threshold values is returned, otherwise the surviving wavelet coefficients are returned

Value

The default output is a list structure, the same length as was input, containing only those wavelet coefficients surviving the threshold.

Details

An extensive amount of literature has been written on wavelet shrinkage. The functions here represent the most basic approaches to the problem of nonparametric function estimation. See the references for further information.

References

Gencay, R., F. Selcuk and B. Whitcher (2001) An Introduction to Wavelets and Other Filtering Methods in Finance and Economics, Academic Press.

Ogden, R. T. (1996) Essential Wavelets for Statistical Applications and Data Analysis, Birkhauser.

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge University Press.

Vidakovic, B. (1999) Statistical Modeling by Wavelets, John Wiley and Sons.

Author

B. Whitcher (some code taken from R. Todd Ogden)