Time-varying and Seasonal Analysis Using Hilbert Wavelet Pairs

Performs time-varying or seasonal coherence and phase anlaysis between two time seris using the maximal-overlap discrete Hilbert wavelet transform (MODHWT).

modhwt.coh(x, y, f.length = 0)

modhwt.phase(x, y, f.length = 0)

modhwt.coh.seasonal(x, y, S = 10, season = 365)

modhwt.phase.seasonal(x, y, season = 365)

Arguments

x

MODHWT object.

y

MODHWT object.

f.length

Length of the rectangular filter.

S

Number of "seasons".

season

Length of the "season".

Value

Time-varying or seasonal coherence and phase between two time series. The coherence estimates are between zero and one, while the phase estimates are between \(-\pi\) and \(\pi\).

Details

The idea of seasonally-varying spectral analysis (SVSA, Madden 1986) is generalized using the MODWT and Hilbert wavelet pairs. For the seasonal case, \(S\) seasons are used to produce a consistent estimate of the coherence and phase. For the non-seasonal case, a simple rectangular (moving-average) filter is applied to the MODHWT coefficients in order to produce consistent estimates.

References

Madden, R.A. (1986). Seasonal variation of the 40–50 day oscillation in the tropics. Journal of the Atmospheric Sciences 43(24), 3138–3158.

Whither, B. and P.F. Craigmile (2004). Multivariate Spectral Analysis Using Hilbert Wavelet Pairs, International Journal of Wavelets, Multiresolution and Information Processing, 2(4), 567–587.

See also

Author

B. Whitcher