Phase Shift Wavelet Coefficients

Wavelet coefficients are circularly shifted by the amount of phase shift induced by the wavelet transform.

phase.shift(z, wf, inv = FALSE)
phase.shift.packet(z, wf, inv = FALSE)

Arguments

z
DWT object

wf
character string; wavelet filter used in DWT

inv
Boolean variable; if inv=TRUE then the inverse phase shift is applied

Value

DWT (DWPT) object with coefficients circularly shifted.

Details

The center-of-energy argument of Hess-Nielsen and Wickerhauser (1996) is used to provide a flexible way to circularly shift wavelet coefficients regardless of the wavelet filter used. The results are not identical to those used by Percival and Walden (2000), but are more flexible.

phase.shift.packet is not yet implemented fully.

References

Hess-Nielsen, N. and M. V. Wickerhauser (1996) Wavelets and time-frequency analysis, Proceedings of the IEEE, 84, No. 4, 523-540.

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge University Press.

Author

B. Whitcher