Plot Two-dimensional Discrete Wavelet Transform

Organizes the wavelet coefficients from a 2D DWT into a single matrix and plots it. The coarser resolutions are nested within the lower-lefthand corner of the image.

# S3 method for dwt.2d
plot(x, cex.axis = 1, plot = TRUE, ...)

Arguments

x
input matrix (image)

cex.axis
par plotting parameter that controls the size of the axis text

plot
if plot = FALSE then the matrix of wavelet coefficients is returned, the default is plot = TRUE

...
additional graphical parameters if necessary

Value

Image plot.

Details

The wavelet coefficients from the DWT object (a list) are reorganized into a single matrix of the same dimension as the original image and the result is plotted.

See also

Author

B. Whitcher