Wavelet-based Maximum Likelihood Estimation for Seasonal Persistent Processes

Parameter estimation for a seasonal persistent (seasonal long-memory) process is performed via maximum likelihood on the wavelet coefficients.

spp.mle(y, wf, J = log(length(y), 2) - 1, p = 0.01, frac = 1)

spp2.mle(y, wf, J = log(length(y), 2) - 1, p = 0.01, dyadic = TRUE, frac = 1)

Arguments

y

Not necessarily dyadic length time series.

wf

Name of the wavelet filter to use in the decomposition. See wave.filter for those wavelet filters available.

J

Depth of the discrete wavelet packet transform.

p

Level of significance for the white noise testing procedure.

frac

Fraction of the time series that should be used in constructing the likelihood function.

dyadic

Logical parameter indicating whether or not the original time series is dyadic in length.

Value

List containing the maximum likelihood estimates (MLEs) of \(\delta\), \(f_G\) and \(\sigma^2\), along with the value of the likelihood for those estimates.

Details

The variance-covariance matrix of the original time series is approximated by its wavelet-based equivalent. A Whittle-type likelihood is then constructed where the sums of squared wavelet coefficients are compared to bandpass filtered version of the true spectral density function. Minimization occurs for the fractional difference parameter \(d\) and the Gegenbauer frequency \(f_G\), while the innovations variance is subsequently estimated.

References

Whitcher, B. (2004) Wavelet-based estimation for seasonal long-memory processes, Technometrics, 46, No. 2, 225-238.

See also

Author

B. Whitcher