Testing for Homogeneity of Variance

A recursive algorithm for detecting and locating multiple variance change points in a sequence of random variables with long-range dependence.

testing.hov(x, wf, J, min.coef=128, debug=FALSE)

Arguments

x
Sequence of observations from a (long memory) time series.

wf
Name of the wavelet filter to use in the decomposition.

J
Specifies the depth of the decomposition. This must be a number less than or equal to \(\log(\mbox{length}(x),2)\).

min.coef
Minimum number of wavelet coefficients for testing purposes. Empirical results suggest that 128 is a reasonable number in order to apply asymptotic critical values.

debug
Boolean variable: if set to TRUE, actions taken by the algorithm are printed to the screen.

Value

Matrix whose columns include (1) the level of the wavelet transform where the variance change occurs, (2) the value of the test statistic, (3) the DWT coefficient where the change point is located, (4) the MODWT coefficient where the change point is located. Note, there is currently no checking that the MODWT is contained within the associated support of the DWT coefficient. This could lead to incorrect estimates of the location of the variance change.

Details

For details see Section 9.6 of Percival and Walden (2000) or Section 7.3 in Gencay, Selcuk and Whitcher (2001).

References

Gencay, R., F. Selcuk and B. Whitcher (2001) An Introduction to Wavelets and Other Filtering Methods in Finance and Economics, Academic Press.

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge University Press.

See also

Author

B. Whitcher